requirements.txt
mysql.connector==2.2.9
inflection==0.3.1